Fix error #50 involving errorPipe, make Stream.whenever more flexible.
errorPipe uses ImmediateLinkUpdate, which will automatically fire off
a handleError event for the bound async.  However, first we must check
to see if an error is pending.  In that case, we can rely on the
upstream async to propagate the error.  We almost must check if the
upstream async is handling the error.  In that case, we don't need to
manage errors on the newly bound async.

Added some helper methods to check for async error handling states.

Also, Stream.whenever now handles generic AsyncBase types correctly in
haxe 3.2.